backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11.5"
dependencies = [
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.4.5",
    "aiohttp==3.11.12",
    "aiosignal==1.3.2",
    "aiosqlite==0.21.0",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "asgiref==3.8.1",
    "attrs==25.1.0",
    "beautifulsoup4==4.13.3",
    "bidict==0.23.1",
    "blinker==1.9.0",
    "cachetools==5.5.1",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "crawl4ai==0.4.248",
    "cryptography==44.0.1",
    "cssselect==1.2.0",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "eventlet==0.39.0",
    "fake-http-header==0.3.5",
    "fake-useragent==2.0.3",
    "fastapi==0.115.8",
    "feedfinder2==0.0.4",
    "feedparser==6.0.11",
    "filelock==3.17.0",
    "filetype==1.2.0",
    "frozenlist==1.5.0",
    "fsspec==2025.2.0",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.24.1",
    "google-api-python-client==2.160.0",
    "google-auth==2.38.0",
    "google-auth-httplib2==0.2.0",
    "google-genai==1.0.0",
    "google-generativeai==0.8.4",
    "googleapis-common-protos==1.66.0",
    "greenlet==3.1.1",
    "grpcio==1.70.0",
    "grpcio-status==1.70.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httplib2==0.22.0",
    "httpx==0.27.2",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.28.1",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "itsdangerous==2.2.0",
    "jieba3k==0.35.1",
    "jinja2==3.1.6",
    "jiter==0.8.2",
    "joblib==1.4.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "langchain-core==0.3.34",
    "langchain-text-splitters==0.3.6",
    "langsmith==0.3.6",
    "litellm==1.61.15",
    "lxml==5.3.0",
    "lxml-html-clean==0.4.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "newspaper3k==0.2.8",
    "nltk==3.9.1",
    "numpy==1.26.4",
    "openai==1.62.0",
    "orjson==3.10.15",
    "outcome==1.3.0.post0",
    "packaging==24.2",
    "pillow==10.4.0",
    "playwright==1.50.0",
    "propcache==0.2.1",
    "proto-plus==1.26.0",
    "protobuf==5.29.3",
    "psutil==6.1.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.7.1",
    "pyee==12.1.1",
    "pygments==2.19.1",
    "pyopenssl==25.0.0",
    "pyparsing==3.2.1",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-engineio==4.11.2",
    "python-socketio==5.12.1",
    "pyyaml==6.0.2",
    "rank-bm25==0.2.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-file==2.1.0",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "rpds-py==0.22.3",
    "rsa==4.9",
    "selenium==4.28.1",
    "sgmllib3k==1.0.0",
    "simple-websocket==1.1.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "snowballstemmer==2.2.0",
    "sortedcontainers==2.4.0",
    "soupsieve==2.6",
    "sqlalchemy==2.0.38",
    "starlette==0.45.3",
    "tenacity==9.0.0",
    "tf-playwright-stealth==1.1.1",
    "tiktoken==0.8.0",
    "tinysegmenter==0.3",
    "tldextract==5.1.3",
    "tokenizers==0.21.0",
    "tqdm==4.67.1",
    "trio==0.28.0",
    "trio-websocket==0.11.1",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "uritemplate==4.1.1",
    "urllib3==2.3.0",
    "uvicorn>=0.34.0",
    "websocket-client==1.8.0",
    "websockets==14.2",
    "werkzeug==3.1.3",
    "wsproto==1.2.0",
    "xxhash==3.5.0",
    "yarl==1.18.3",
    "zipp==3.21.0",
    "zstandard==0.23.0",
]
